chore: rename `widthdraw` to `withdraw`
