+++ lib/libs/email/content/index.ts
@@ -1,7 +1,7 @@
 export * from "./new-submission";
 export * from "./tempExtension";
 export * from "./respondToRai";
-export * from "./widthdrawRai";
+export * from "./withdrawRai";
 export * from "./withdrawPackage";
 export * from "./email-components";
 export * from "./respondToRai";
